Put Ctrl-C checking in mid-evaluation
The previous checking for Ctrl-C tried to only run on completed value
fulfillment.  However, this means you wouldn't get the check opportunity
if things went into a never-ending opening of frames without closing.

This moves the checking to the middle of evaluations, pending any
better ideas.  That means it is checked more frequently than it might
need to be, so the eval tick count before running the actual check
might need to be adjusted.