Begin stackless TRAP; separate global state from longjmp
This lays the foundation for a stackless form of trap, in which the
single trampoline needs only one `setjmp` and cpu state in order to
implement any number of TRAPs (or traplike behaviors).  This contrasts
with the non-stackless approach, where each TRAP had to set up its own
setjmp in order to have an opportunity to add an interception point
for the longjmp.

Hence the 1-to-1 relationship between setjmp buffers and TRAPs no
longer existed.  So the rollback information about mold buffers and
data stack was kept in Reb_State (needing at least one per TRAP) and
the jump buffer into its own Reb_Jump structure.

(Though a pure stackless Rebol system would not require more than one
jmp_buf, not all natives are pure (yet)...and also, any callouts to
non-Rebol code which wishes to re-enter without giving up the stack
would need to invoke another trampoline.  So there still needs to be
a list of where those jump buffers are, necessitating Reb_Jump).

Once the properties were separated out, the idea of what a Reb_State
was truly capturing became clear as something each frame might
potentially need to know.  Not only did TRAP need it, but the
YIELD would need it to know how much of the mold buffer to account
for (previously only adjusted for stack states).

So this moves toward generalizing the notion of what needs to be
rolled back from global shared state into a "baseline", which is a
Reb_State that is in some way memoized in each stack frame.  That
recognition of the related group of concerns motivated factoring out
the state-related rollback, unplugging, and replugging routines into
a separate %c-state.c module.

Because many of the comments describing the setjmp() and longjmp()
behavior predated understanding of what R3-Alpha was doing, this
was an opportunity to make them more clear.  As one comment notes,
this is a work in progress--e.g. currently the only global state
handled by YIELD's unplugging and replugging is the data stack (so
mold buffers will not survive if an in progress mold is interrupted).