Handle mold buffer in Unplug_Stack(), stackless DELIMIT
This adds the mold buffer to the things that are able to be handled
across a YIELD.  The mechanism used involves array flags on the
data stack fragment that was already there, indicating that some values
on the top of the "stack" are actually more fragments...e.g. that the
top is actually a TEXT! value of the mold buffer fragment.  This
avoids creating more costly objects.

To test the new facility, DELIMIT is converted to stackless.  This
means that SPACED, UNSPACED, etc. also become stackless (!).