Simpler frame phase dispatch, deprecate Dispatch_Internal()
A concept used to start developing usermode debugging and tracing
constructs was that they would be able to "hook" the central function
calling mechanics by wrapping them.  This would be somewhat similar to
an ENCLOSE, where the frame was available for each function call to
inspect.  The idea of doing this hooking by replacing C functions and
then calling C functions is intrinsically flawed in a stackless model,
though the code for the old demos depending on the technique is not
deleted yet...as it might be helpful in rethinking how the demos would
work more generally in a stackless model.  Having a stub for this
routine around keeps the broken code compiling.

Also, putting the dispatcher into a separate variable and calling the
function directly makes stepping in significantly easier (as there are
indirections to get to that phase, due to hijackability and other
reasons.)