Use Abort_Frame() on THROW to process more globals
Prior to stackless processing, all calls into the evaluator would
return a flag as to whether there was a throw during the operation that
needed to be passed on.  The routine that made the call could then
synchronously undo any aggregated global state that had been accrued.

An aspect of stackless is that common global state (mold buffer, data
stack) can be interrupted at any time and then put back.  To lighten
the burden on natives from having to have some kind of serialization
callback, this is taken care of automatically.

However, the data stack was dropped to "baseline" during a throw
without dropping the mold stack.  This changes the call from
Drop_Frame() to Abort_Frame()--the same call used in error handling.
That accounts for the mold stack too.