SET-WORD!/PATH! stackless, use STATE_BYTE() in evaluator
There was no "state byte" chewed out of the evaluator flags in the
original stackless implementation.  Questions of granularity remain
about what steps are visible to hook in a debugger or otherwise, but
this switches to a mode of using a state byte instead of a separate
frame "Executor" for group evaluations.

With the state byte available, it is used to allow the SET-WORD!,
SET-PATH!, and SET-GROUP! cases to be stackless.  This tampers with
some optimization that was being made in the feed to sometimes not
use a new frame, but that performance tweak is saved for another time.