Adjust mold baseline for nested frame during DELIMIT
The DELIMIT operation holds a subframe alive for each evaluation step,
but that subframe was not having its baseline concept of where the
mold buffer was adjusted at each step.  So each nested delimit was
thinking it was supposed to pop the entire mold contents when it
completed.

This does a manual adjustment of the baseline--which is awkward, but
necessary.  Better mechanisms may need to emerge if this pattern turns
out to be common.