Initial factoring out of `functionals` directory
In the interests of clarity, this bundles together various "derived
function generators" with their dispatchers into single files which
can explain their purpose and implementation.  (Previously these
were split into a file that contained the dispatchers, and another
that contained the natives which set up the functions with those
dispatchers.)

More work needs to be done on the organization and naming, but this is
a step in the right direction for making it easier for people to see
the pattern and how new "function generators" might be created.