Cleanup WAIT, Reduce_To_Stack_Throws() out of Core
The WAIT native was built on cryptic functions (Wait_Ports_Throws(),
Sieve_Ports(), Awake_System()) that were called *only* from WAIT.
This commit starts by folding those functions into a clearer single
implementation showing what the routine does (and didn't do that it
thought it was).

One thing that it did was call Reduce_To_Stack_Throws() on a block it
received so you could write something like:

     timeout: 1:00
     wait [timeout]  ; block is reduced so it sees the 1:00

In "stackless" code, one tries to avoid calling routines that will
recurse the evaluator.  By its definition, the Reduce_To_Stack_Throws()
function needs to make several evaluator calls...pushing each result
to the stack, before returning.

One very good way to avoid this pattern in C routines implementing
natives is to do as much pre-processing of inputs as possible in
usermode.  This approach is taken with the WAIT operation, which is
broken down into a WAIT* which assumes its block input has already
been reduced...and a higher-level WAIT that does the REDUCE for it.

This was not an option for an experimental feature in PARSE which
let you reduce a BLOCK! for KEEP input.  This may not be desirable,
but it was restyled to use a CONTINUE_WITH () continuation that
called the REDUCE native with the block to reduce.

Consequently, the only remaining piece of code that is using the
Reduce_To_Stack_Throws() is code in the FFI.  The FFI was originally
made by Atronix, and was written in C before good concepts of dividing
labor in extenssions between usermode and native code were established.
This moves the code there pending a rewrite of MAKE STRUCT! which would
better break things down to use far less C than it does today.