Fix REDUCE problem when reducing non-BLOCK!s and GROUP!s
The CONTINUE_WITH and DELEGATE_WITH operations were changed to be more
efficient by actively putting their argument in the First_Unpecialized
slot.  This does not play well with variadics, which expect to be
taking values from a frame...and the nulled state of the variadic
causes problems.

(This needs to be reviewed in the future, if variadics are intended to
work with the CONTINUE_WITH mechanic.)

In any case, it broke the way that REDUCE was processing non-BLOCK!
and non-GROUP! arguments to call REEVAL.  This basically changes it
to repeat the code from REEVAL instead--which is more efficient and
actually not much code to repeat.