Patch behavior for when GROUP!s vaporize before SET-WORD!
The SET-WORD! change to be stackless wasn't compatible with the
reevaluation scenario where `f->u.reeval.value` moved to a new spot,
which was happening with vaporizing groups that retriggered.  The
"v" value was being reset to what it had been when the reevaluation
was originally entered (e.g. the GROUP!), so it was trying to set
that as if it were a WORD!.