Fledgling new unwinding model for fail() errors
This is a somewhat messy patchwork of changes, implementing a start
on a design that leverages the fact that the Trampoline() function
can use a single setjmp() to protect an arbitrary number of Revolt
stack levels.

Before that was the case, only an explicit (and costly) invocation
of a native like TRAP or other barrier could catch an "abrupt"
failure.  These were exception-like moments that would interrupt the
control flow of the C code...not making it to a `return`, and an
arbitrary number of Revolt stack frames which were actively in mid
execution would need to have the rug pulled out from under them.
Some amount of cleanup was done in a stack walk prior to doing the
longjmp, but this cleanup was on frames that were making a call
into the evaluator.

With this new model, only the frame that was executing during the
fail() is interrupted.  It is tagged with ABRUPT_FAILURE, and the
"Executor" function for the frame gets a chance to react.  This is
helpful for intance when the Action_Executor() delegates to a
native's "Dispatcher" function which fails...that failure is still
under the umbrella of the Action_Executor()'s frame, but it gets
a chance to drop the varlist for the frame so it does not leak.
This places the concern for how that kind of frame is cleaned up
to be the responsibility of the executor.

Once the ABRUPT_FAILURE frame has been given its chance to react,
the error is then converted into a THROW-style cooperative failure.
It is passed up the stack and managed like any other throw is, in
a non-ABRUPT way.  If it reaches the edge of a trampoline on a
residual stacklike-call, then it will fail() again, leading to
one frame of ABRUPT_FAILURE above it again.

Throwing a wrench into this model is the desire to be able to target
evalutions into cells other than a frame's "out" pointer.  This makes
ABRUPT_FAILURE different from a throw, because the Trampoline pokes
the failure into the output cell...which isn't where it would be
looking for a throw that was bubbling up through a different target
location.  This suggests that perhaps ABRUPT_FAILURE needs a special
handling mechanism...but for now it is tied into putting the failure
in the frame's out cell, because in the average case that can be
handled in the same way a throw from a frame beneath would be if
`f->out` was the target.

It's a first pass which adds a number of asserts to try and push a bit
of the complex logic out of the Trampoline and into the "Executors"
where they have more knowledge of what needs to be done.  The details
still need more work...but it seems a good direction, especially with
how legible the Trampoline is becoming.