Simplify PUSH_TRAP() to just take a jump buffer
This elongates the name of PUSH_TRAP() to more clearly convey its
"supernatural" setjmp() wrapping character.  It also goes ahead and
has the callsites reference the error field that is stored inside of
the jump buffer, to reduce it down to taking just one parameter.

(The idea of letting the caller provide the address of an error to
write to originated in a time when PUSH_TRAP() calls were more
numerous, and it could be common that someone might have an error
variable they wanted to set that already existed.)

The amount of code that is entangled with the setjmp logic had dwindled
to just three places: rebRescue(), rebRescueWith() and the trampoline.
In order to pare this down to just two places this commit implements
rebRescue() in terms of rebRescueWith() where the rescuing handler is
simply nullptr.  In doing so it also updates the implementation of
rebRescueWith() to be correct, because it had not been used and thus
hadn't been brought up to date.