First-cut hacky GC for non-terminated generators
A generator may not run to completion, and may not even have a
completed state.  e.g. this generator is infinite:

    g: generator [while [true] [yield <1020> yield #0304]]

Their suspended state consists of detached REBFRM* stacks between
the generator frame and the last yield that ran.  This state had no
previous mechanism for freeing, which meant they showed up as leaks
on shutdown.

This commit does rearranging to move closer to having an answer,
by walking up the detached frame stack.  It leverages the invariant
that was added where STATE_BYTE() of a frame is never zero for a
continuation, to help enforce a proper order of disposal for frames
where the lower frames are freed before the higher ones.

(Better answers would delegate to the Executor of the frames to do
the freeing, allowing them to cleanly abort in case they allocated
state.  But this makes it possible for common cases to shut down
cleanly, for now.)