Enforce STATE_BYTE() != 0 for R_CONTINUATION pushes
Before STATE_BYTE() was chewed out of the frame's flags, there was the
idea of using the FRM_SPARE() cell to store information about which
phase of execution a native or dispatcher's "state machine" was in.
This wound up taking 4 platform pointers of usable GC-protected cell
space to do what a byte could do...and being able to encode that byte
in the "hot" platform-pointer-sized frame flags was better (especially
since it could be initialized or zeroed out at the same time all the
other flags were being set in the frame).

A concept arose for being able to tell whether a frame was in a
continuation state or not.  Because the frame stack is a singly linked
list from deeper stack frames up to shallower ones, it's not possible
to tell by looking at a pointer.

Rather than introduce a flag for "this frame has a continuation in
process", the notion of forcing all continuations to happen with a
nonzero state byte seemed like a reasonable idea.  This wouldn't be
putting too much of an inconvenience on most stackless routines,
which already were using nonzero state bytes anyway.

Some cases that don't care (like DELEGATE()) are inelegantly set here
to just assign 1.  This should be revisited, as CONTINUE() probably
should be taking a state byte argument anyway...and R_CONTINUATION
shouldn't be used directly.