Add error and throw robustness to Yielder/Generator
Errors and Throws interrupt and unwind the stack in a non-recoverable
way.  So if these reach the top level of a yielder without being
CATCH'd or TRAP'd, then the yielder cannot be run again.

This makes it so that a failure that reaches the yielder will make it
so that it returns an error saying that it failed.  However, a throw
that passes the yielder is considered as valid as any other form of
exiting control flow (e.g. RETURN) and just makes the yielder stop.

It also introduces the first instance of a cooperative failure, where
a failure is returned through the normal throwing channels vs. causing
an "abrupt" failure that has to be caught by the trampoline.  This is
slightly more efficient and may be something to consider doing more
often (or not, if it makes the source seem more complicated).