Save 5424 bytes by storing REBACT* for natives
The table of natives created was using REBVAL cells, when REBACT*
would suffice.  (The design of ACTION! is such that not only can
the paramlist be acquired from a cell, but an "archetype" cell for
the action sits in the [0] slot of the paramlist and can be returned
as a value.)

This saves 3 platform pointers per native.

It also changes NAT_ACTION and NAT_VALUE to NATIVE_VAL and NATIVE_ACT,
in anticipation of GENERIC_VAL and GENERIC_ACT (which would be less
ambiguous than GEN_VALUE and GEN_ACTION that look like "generate").