Attempt to make console itself stackless
The initial idea with GO routines was to keep the main execution
thread distinct, so it could remain on the stack at all times.  This
permitted some amount of stackful frames...so long as they were above
any potential YIELD or other unwind points, at the top of the chain.

In the interests of uniformity, this experiment tried to move away
from that and make the console stackless.  That might be something
that has to be done :-/ but for the moment it means not only does
console need to be stackless, but the call from main() to invoke the
console would have to be stackless too.  This opens cans of worms
that aren't necessarily super easy to reason about...putting aside
for the current moment.