Initial Roaring Bitmaps Implementation
This commit seeks to attack the countless problems with R3-Alpha's
BITSET! datatype.  Instead of using a thin veneer over BINARY!, this
links in the Roaring Bitmaps pure C library:

https://roaringbitmap.org/

In order to avoid duplicating the allocation of a tracking entity for
the bitmap and a separate series node to use for GC, some slight
surgery is done on the roaring bitmap library so that it can use a
REBSER node as the API handle.  This requires calls that want to use
the bitset to unpack the node into an API structure's layout, and
a few hooks are injected into the roaring functions so that they keep
the series node in sync with certain changes.

Some issues are punted on for the moment...such as how to mold the
bitmaps.  Only the first few numbers are shown as integers (as
a bitset of a single number negated contains all integers but that).

Roaring offers many more bitset features than R3-Alpha, but this does
not expose those.  It seeks only to do what was covered by boot and
by testing (though there was very minimal testing!)