Words sacrifice spelling when bound (WIP)
This switches it so that a WORD! will give up its spelling once it
is bound, using a dereference to the bound context to get it.  This
means getting the spelling costs slightly more, but it frees up a
slot in the word for other uses.  That use is intended to be a caching
strategy for virtual binding.

PROBLEM: This does not preserve the *specific* case variation of the
word used in the bind.  That could be done by chewing out a few bits
from the index field.  Or the cells could just be made bigger...but
this is still trying to preserve the "sweet spot" of 4 pointers.  A
solution is punted on for the moment, while using the opened-up space
is tried as proof of concept.