Split VAL_CONTEXT_BINDING() from VAL_BINDING()
This does the same split out of context bindings that the previous
commit did for action bindings.  Notably, only frame contexts can
have bindings...which are caches of the bindings of the actions to
which they are attached (so that e.g. MAKE FRAME! of a RETURN knows
which function instance to return from).