Systemize rootvar initialization
The initialization of rootvars had evolved "organically" and had
a fair amount of redundancy.  The desire to get the same protection
from CELL_FLAG_PROTECTED that actions have on their archetype
motivated centralizing the initialization somewhat.

Since these root initializations make it clear in one place where
non-frame cells are setting binding and phase to unused pointers,
this allows other forms of initialization to make it clear that they
are only intended to work on FRAME! values.